Chemistry.Bond

Description

A module that defines the different types of chemical bonds together with conversions from and to corresponding unicode characters.

Types of bonds

data Bond

A datatype that defines the different types of bonds between atoms. There are single, double, triple and quadruple bonds. Bond is an instance of Monoid where we implement it by picking the maximum bond. The neutral element is thus a single bond.

Constructors

BSingle
A single bond.

BDouble
A double bond.

BTriple
A triple bond.

BQuadruple
A quadruple bond.

Rendering and parsing bonds

bondToUnicode :: Bond -> Char

Convert the given Bond type to a unicode character to format that binding.

bondFromUnicode :: Char -> Maybe Bond

Parse the given Character to the corresponding Bond. It makes use of the characters defined in bondToUnicode as well as characters defined by the SMILES standard.

bondChars :: [(Char, Bond)]

A list of 2-tuples where the left item defines a Character and the right item the corresponding Bond. This is used for parsing purposes.