Chemistry.Charge

Description

A module that defines datatypes to specify the charge of an atoms.

Documentation

data Charged a

A data type that specifies that the given item has a charge.

Constructors

Charged a Int
A dataconstructor that defines a Charged item together with the charge.

neutral

Arguments

:: a
The given item to wrap in a Charged data constructor.

-> Charged a
The corresponding Charged item of the given chemical item.

Construct a Charged item with charge 0.