Chemistry.Core

Description

A module that defines the main typeclasses and functions for parsing, presenting and rendering molecules.

A collection where items can be multiplied

class QuantifiedElements f where

A typeclass that specifies a collection where items can have a multiplication factor.

Minimal complete definition

foldQuantified

Methods

foldQuantified

Arguments

:: (a -> b)
The given mapping function that maps items to a result type.

-> (b -> b -> b)
The given append function that is triggered when we have a combination of two QuantifiedElements subexpressions.

-> (Int -> b -> b)
The given multiplication function when a certain subexpression occurs a given number of times.

-> f a
The QuantifiedElements object that will be folded.

-> b
The corresponding result of the fold operation.

Fold a given QuantifiedElements object by defining rules for mapping, appending, and multiplying.

listElements'

Arguments

:: f a
The given QuantifiedElements to determine the items from.

-> [(a, Int)]
The given list to append to the result.

-> [(a, Int)]
The result of determining the number of occurrences of the chemical items together with the given list.

List the elements in the QuantifiedElements expression and append a given list. Items can occur multiple times.

listElements

Arguments

:: f a
The given QuantifiedElements to determine the item count from.

-> [(a, Int)]
A list of items with the number of times these occur. An item can occur multiple times.

List the elements in the QuantifiedElements. Items can occur multiple times.

listElementsCounter

Arguments

:: (Eq a, Hashable a)

=> f a
The given QuantifiedElements to determine the item count from.

-> HashMap a Int
A HashMap that maps the items to the number of times these occur.

Convert the given QuantifiedElements to a HashMap that maps each item to the number of times it is present.

Rendering formulas as text and HTML

class FormulaElement a where

A typeclass to represent types that can be rendered as chemical formulas.

Minimal complete definition

toFormulaPrec | toFormula

Methods

toFormula

Arguments

:: a
The given item to render.

-> Text
A Text object that represents the given chemical item.

Render the given item as Text.

toFormulaPrec

Arguments

:: Int
The given precedence value.

-> a
The chemical item to render.

-> Text
The Text object that is the "tail" of the Text that will be generated.

-> Text
The Text that represents the chemical item appended with the given tail.

Render the given chemical item with the precedence value.

toFormulaMarkup

Arguments

:: a
The given item to render.

-> Markup
A Markup object that represents the given chemical item.

Render the given chemical item as Markup.

toFormulaMarkupPrec

Arguments

:: Int
The given precedence value.

-> a
The chemical item to render.

-> Markup
The given Markup object that is the "tail" of the Markup that will be generated.

-> Markup
The Markup that represents the chemical item appended with the given tail.

Render the given chemical item with the precedence value.

Items that have weight

class Weight a where

A typeclass that represents that the given type has a weight.

Methods

weight

Arguments

:: Floating b

=> a
The given item to determine the weight from.

-> Maybe (Quantity DMass b)
The weight of the given item, given the item has a weight.

Calculate the weight of the given item.

quantifiedWeight

Arguments

:: (QuantifiedElements f, Weight a, Floating b)

=> f a
The given item to determine the weight from.

-> Maybe (Quantity DMass b)
The corresponding weight of the given checmical item.

Determine the weight of the given QuantifiedElements object. This is implemented by summing and multiplying the weights of the individual elements.

Hill system

class HillCompare a where

A typeclass that spans an order relation based on the rules of a Hill system.

Methods

hillCompare

Arguments

:: a
The first item to compare with the Hill system.

-> a
The second item to compare with the Hill system.

-> Ordering
The result of the comparison with the Hill system.

Compare the two given items according to the hill system.

Helper functions to render parenthesis

showParenText

Arguments

:: Bool
A Bool that determines if we generate parenthesis.

-> (Text -> Text)
A text generator function that can be wrapped in parenthesis.

-> Text

-> Text
The resulting text generator.

A helper function to wrap the text generator in parenthesis if the given condition is True, and simply returns the text generator otherwise.

showParenMarkup

Arguments

:: Bool
A Bool that determines if we generate parenthesis.

-> (Markup -> Markup)
A markup generator function that can be wrapped in parenthesis.

-> Markup

-> Markup
The resulting markup generator.

A helper function to wrap the markup generator in parenthesis if the given condition is True, and simply returns the markup generator otherwise.