Chemistry.Formula

Description

A module that defines datastructure to present formulas and linear chains.

Representing formulas

data FormulaPart a

A data type to specify the part of a formula. This allow to multiply a Formula, but not concatenate.

Constructors

FormulaItem a
Wrap an item to present it as a FormulaPart.

(Formula a) :* Int infix 8
Multiply a Formula with a given number.

data Formula a

A data type to specify a formula. This allows to concatenate FormulaParts together, but not multiply these.

Constructors

FormulaPart (FormulaPart a)
Wrap a FormulaPart to present it as a Formula.

(FormulaPart a) :- (Formula a) infixr 7
Prepend a Formula with a FormulaPart.

(.*) infixr 8

Arguments

:: FormulaPart a
The original FormulaPart.

-> Int
The given multiplier.

-> FormulaPart a
The given FormulaPart multiplied with the given multiplier.

Multiply the given FormulaPart with a given number to a new FormulaPart.

formulaToParts

Arguments

:: Formula a
The given formula to convert.

-> NonEmpty (FormulaPart a)
A NonEmpty list of FormulaParts the Formula consists out of.

Convert a given formula to a NonEmpty list of FormulaParts.

fromElementList

Arguments

:: [(a, Int)]
The given list of elements with the corresponding quantity.

-> Formula a
The corresponding Formula.

Convert a given list of items with their quantity to a Formula. This does not allow to multiply a "formulapart" like C2H4(OH)2.

Conversion to canonical forms

toMolecular :: (Eq a, Hashable a) => Formula a -> Formula a

Convert the given Formula to the a Formula that is a molecular formula. In this formula, each atom is denoted once with the total number of times it occurs.

toHillFormula :: (Eq a, Hashable a, HillCompare a) => Formula a -> Formula a

Convert the given Formula to a Formula that presents the molecule with the Hill system. In this system carbon is denoted first followed by hydrogen, etc.

Representing linear chain

data LinearChain bond element

A data type to present a chain of elements that are separated by (possibly different) bond.

Constructors

ChainItem element
Wrap a single element into a LinearChain.

Chain element bond (LinearChain bond element)
Prepend an existing element with an element through a binding.

(.-) infixr 6

Arguments

:: element
The element to prepend to the LinearChain.

-> LinearChain Bond element
The rest of the LinearChain.

-> LinearChain Bond element
A LinearChain where the given element is prepended through a single bond.

Create a LinearChain by prepending an element with a single bond to a LinearChain.

(.=) infixr 6

Arguments

:: element
The element to prepend to the LinearChain.

-> LinearChain Bond element
The rest of the LinearChain.

-> LinearChain Bond element
A LinearChain where the given element is prepended through a double bond.

Create a LinearChain by prepending an element with a double bond to a LinearChain.

(.#) infixr 6

Arguments

:: element
The element to prepend to the LinearChain.

-> LinearChain Bond element
The rest of the LinearChain.

-> LinearChain Bond element
A LinearChain where the given element is prepended through a triple bond.

Create a LinearChain by prepending an element with a triple bond to a LinearChain.

(.$) infixr 6

Arguments

:: element
The element to prepend to the LinearChain.

-> LinearChain Bond element
The rest of the LinearChain.

-> LinearChain Bond element
A LinearChain where the given element is prepended through a quadruple bond.

Create a LinearChain by prepending an element with a quadruple bond to a LinearChain.