Chemistry.Isotope

Description

A module that defines datatypes to specify the isotope of an atoms.

Representing an isotope

data Isotope a

A datatype that specifies that an item has a specific isotope.

Constructors

Isotope
A dataconstructor that contains the item together with its mass number.

Fields

isoElement :: a

The chemical item for which we work with an isotope.

massNumber :: Int

The mass number of that chemical element.

Weight of isotopes

isotopeWeight

Arguments

:: Floating a

=> Element
The given element for which we determine the weight.

-> Int
The given mass number of the element for which we determine the weight.

-> Maybe (Quantity DMass a)
The corresponding weight of the given Element with the given mass number.

Obtain the weight of the given isotope. For isotopes that are impossible or unknown, Nothing is returned.