Chemistry.Parser

Description

A module that defines parsers to parse elements, formulas, bonds, etc.

Parsing elements

elementParser

Arguments

:: Stream s m Char

=> ParsecT s u m Element
The parser that reads to an Element.

A parser that aims to convert text into an Element.

Parsing bonds

bondParser

Arguments

:: Stream s m Char

=> ParsecT s u m Bond
A parser to parse the different types of bonds.

A parser that parses the different types of bonds as defined by the bondChars.

Parsing charged items

chargedParser

Arguments

:: Stream s m Char

=> ParsecT s u m (Charged Element)
The parser that parses charged Elements.

A parser that parses charged items. The charge can be represented with a plus or minus followed by a number, or a sequence of plusses and minusses where the number of characters determine the magnitude of the charge.

chargedParser'

Arguments

:: Stream s m Char

=> ParsecT s u m a
The given parser for the chemical item.

-> ParsecT s u m (Charged a)
A parser that will parse the chemical item followed by the charge.

A parser that uses a parser, and reads that parser together with the charge. The charge can be represented with a plus or minus followed by a number, or a sequence of plusses and minusses where the number of characters determine the magnitude of the charge.

Parsing elements with quantities

quantityParser

Arguments

:: Stream s m Char

=> ParsecT s u m a
The element parser that is used.

-> ParsecT s u m (a, Int)
A parser that will parse the element together with an optional quantity.

A function that is given a parser to parse an item and produces a parser that parses the element with its quantity. If no quantity is given, one is used.

elementQuantityParser

Arguments

:: Stream s m Char

=> ParsecT s u m (Element, Int)
The parser that parses a combination of an element and a quantity.

A parser that parses an Element together with an optional quantity. if no quantity is passed, one is used.

Parsing formulas

formulaParser

Arguments

:: Stream s m Char

=> ParsecT s u m (Formula Element)
A parser that parses Formulas with Elements as items.

A parser that parses Formulas with Elements as items.

formulaParser'

Arguments

:: Stream s m Char

=> ParsecT s u m a
The given parser that parses an element.

-> ParsecT s u m (Formula a)
The resulting parser that parses formulas with elements parsed by the given element parser.

A parser that will for a given element parser parse a formula which is a combination of concatenations and multiplications.

chargedFormulaParser

Arguments

:: Stream s m Char

=> ParsecT s u m (Formula (Charged Element))
A parser that parses Formulas with Elements that can contain a charge.

A parser that parses Formulas with Elements that can contain a charge.

Parsing linear chains

linearChainParser

Arguments

:: Stream s m Char

=> ParsecT s u m (LinearChain Bond (Formula Element))
A parser that parses a LinearChain of Bonds and Formulas of Elements

A parser that parses a LinearChain of Bonds and Formulas of Elements. This makes use of the linearChainParser'', and uses other parsers to parse the bond and formula.

linearChainParser'

Arguments

:: Stream s m Char

=> ParsecT s u m a
The parser for the items that will be used.

-> ParsecT s u m (LinearChain Bond a)
A parser that parses LinearChains with the given parser for the items.

A function that produces a parser that parses a LinearChain that parses Bonds and makes use of a given parser to parse the items in the linear chain.

linearChainParser''

Arguments

:: Stream s m t

=> ParsecT s u m bond
The given parser to parse the bonds.

-> ParsecT s u m element
The given parser to parse the items.

-> ParsecT s u m (LinearChain bond element)
A parser that parses LinearChains for the given bond and item parsers.

A function that produces a parser that parses a LinearChain that makes use of given parsers to parse the bonds and the items in the linear chain.

chargedLinearChainParser

Arguments

:: Stream s m Char

=> ParsecT s u m (LinearChain Bond (Formula (Charged Element)))
A parser that parses a LinearChain of Bonds and charged Formulas of Elements

A parser that parses a LinearChain of Bonds and Formulas of charged Elements. This makes use of the linearChainParser'', and uses other parsers to parse the bond and formula of charged elements.

Quasiquoters

elqq

Arguments

:: QuasiQuoter
A quasiquoter to parse Elements.

A quasiquoter that can parse chemical Elements. The quasiquoter can be used for expressions and patterns. If it is used as a type signature, it will be replaced with the Element type.

elqqq

Arguments

:: QuasiQuoter
A quasiquoter to parse a combination of an Element and its quantity.

A quasiquoter that can parse chemical Elements with an optional quantity. The quasiquoter can be used for expressions and patterns. If it is used as a type signature, it will be replaced with a 2-tuple type with an Element and an Int.

chelqq

Arguments

:: QuasiQuoter
A quasiquoter that parses charged Elements.

A quasiquoter that can parse Charged chemical Elements with an optional quantity. The quasiquoter can be used for expressions and patterns. If it is used as a type signature, it will be replaced with a Charged Element.

formulaqq

Arguments

:: QuasiQuoter
A quasiquoter that parses Formulas of Elements.

A quasiquoter that can parse Formulas of Elements. The quasiquoter can be used for expressions and patterns. If it is used as a type signature, it will be replaced with a Formula Element.

chainqq

Arguments

:: QuasiQuoter
A quasiquoter that parses LinearChains of Bonds and Formulas of Elements.

A quasiquoter that can parse a LinearChain of Bonds and Elements. The quasiquoter can be used for expressions and patterns. If it is used as a type signature, it will be replaced with a LinearChain of Bonds and Elements.

chchainqq :: QuasiQuoter

A quasiquoter that can parse a LinearChain of Bonds and Charged Elements. The quasiquoter can be used for expressions and patterns. If it is used as a type signature, it will be replaced with a LinearChain of Bonds and Charged Elements.